requirements.txt
celery
pymongo>=3
requests
flower
folium
bearcart
bokeh
psycopg2
sqlsoup
odo
geojson
python-dotenv
dbfread
pandasql
python-metar
pytest
PyYAML
